Moved start/stop_services.sh from app/example to common

+++ common/stop_services.sh
@@ -2,5 +2,7 @@
 webrun_directory=[==WEBRUN_DIRECTORY==]
 continue_upload_monitor=$webrun_directory/upl/CONTINUE_UPLOAD_MONITOR
 continue_xml_import=$webrun_directory/CONTINUE_XML_IMPORT
-rm -f $continue_upload_monitor
+if [ -d $webrun_directory/upl ]; then
+   rm -f $continue_upload_monitor
+fi
 rm -f $continue_xml_import
